# api/swagger/rest/v1/cluster-service.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Cluster service",
    "version": "1.0",
    "contact": {
      "name": "Cluster service",
      "email": "trifan@adobe.com"
    }
  },
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/clusters": {
      "post": {
        "summary": "Create new clusters",
        "operationId": "CreateCluster",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreateResponseCluster"
            }
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "format": "string"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateRequestCluster"
            }
          }
        ],
        "tags": [
          "ClusterService"
        ]
      }
    },
    "/v1/clusters/{nodeId}": {
      "get": {
        "summary": "Read all clusters for a node",
        "operationId": "ReadAllClustersForNode",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ReadAllResponseForNodeCluster"
            }
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "format": "string"
            }
          }
        },
        "parameters": [
          {
            "name": "nodeId",
            "description": "API versioning: it is my best practice to specify version explicitly",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "ClusterService"
        ]
      },
      "delete": {
        "summary": "Delete cluster from node",
        "operationId": "DeleteCluster",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeleteResponseCluster"
            }
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "format": "string"
            }
          }
        },
        "parameters": [
          {
            "name": "nodeId",
            "description": "node id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "versionId",
            "description": "version of cluster to delete.",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "ClusterService"
        ]
      }
    }
  },
  "definitions": {
    "v1Cluster": {
      "type": "object",
      "properties": {
        "clusterName": {
          "type": "string",
          "title": "Unique clusterName"
        },
        "LbPolicy": {
          "type": "string",
          "title": "Type of lb policy"
        },
        "hosts": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1Hosts"
          },
          "title": "list of hosts"
        }
      },
      "title": "Cluster message\nlast id 4"
    },
    "v1CreateRequestCluster": {
      "type": "object",
      "properties": {
        "nodeId": {
          "type": "string",
          "title": "for which node to create the cluster"
        },
        "cluster": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1Cluster"
          },
          "title": "List of clusters"
        },
        "version": {
          "type": "string",
          "title": "the version"
        }
      },
      "title": "Request data to create new Cluster"
    },
    "v1CreateResponseCluster": {
      "type": "object",
      "properties": {
        "nodeId": {
          "type": "string",
          "title": "The node where the cluster has been created"
        },
        "version": {
          "type": "string",
          "title": "versions of the created clusters"
        }
      },
      "title": "Contains data of created Cluster\nlast id 2"
    },
    "v1DeleteResponseCluster": {
      "type": "object",
      "properties": {
        "nodeId": {
          "type": "string",
          "title": "nodeId response"
        },
        "deleted": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "title": "contains the versions of deleted clusters"
        }
      },
      "title": "Contains status of delete operation"
    },
    "v1Hosts": {
      "type": "object",
      "properties": {
        "ip": {
          "type": "string"
        },
        "port": {
          "type": "integer",
          "format": "int64"
        }
      },
      "title": "Hosts message\nlast id = 2"
    },
    "v1ReadAllResponseForNodeCluster": {
      "type": "object",
      "properties": {
        "nodeId": {
          "type": "string",
          "title": "API versioning: it is my best practice to specify version explicitly"
        },
        "clusters": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1Cluster"
          },
          "title": "List of all todo tasks"
        }
      },
      "title": "Contains list of all clusters on a node"
    }
  }
}